+++ CHANGELOG.rst
@@ -14,6 +14,7 @@
 
   Also fixed ``init_popen_io`` to use ``closefd=False`` for shared stdin and stdout file
   descriptors, preventing ``Bad file descriptor`` errors triggered by test_stdouterrin_setnull.
+* Fixed ``GatewayBase.join()`` timeout argument getting ignored.
 * Removed support for Python 3.7.
 * Added official support for Python 3.12.