+++ CHANGELOG.rst
@@ -15,8 +15,8 @@
   Also fixed ``init_popen_io`` to use ``closefd=False`` for shared stdin and stdout file
   descriptors, preventing ``Bad file descriptor`` errors triggered by test_stdouterrin_setnull.
 * The library is now typed and the typing is exposed to type-checkers.
-* Re-exported ``Gateway`` and ``Channel`` from ``execnet``. The constructors
-  are private.
+* Re-exported ``Gateway``, ``Channel``, ``DumpError`` and ``LoadError`` from
+  ``execnet``. The constructors are private.
 * Fixed ``GatewayBase.join()`` timeout argument getting ignored.
 * Removed support for Python 3.7.
 * Added official support for Python 3.12.